# services/auth/cmd/api/handlers/token.go
package handlers

import (
	"encoding/json"
	"net/http"
)

// VerifyAccessToken verifies an access token
// @Summary Verify access token
// @Description Verifies if the provided access token is valid
// @Tags XFWToken
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param token query string false "Access token to verify"
// @Success 200 {object} VerifyAccessTokenResponse
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Router /api/v1/XFWToken/verify-access-token [get]
func VerifyAccessToken(w http.ResponseWriter, r *http.Request) {
	// Get token from query parameter or Authorization header
	token := r.URL.Query().Get("token")
	if token == "" {
		authHeader := r.Header.Get("Authorization")
		if len(authHeader) > 7 && authHeader[:7] == "Bearer " {
			token = authHeader[7:]
		}
	}

	if token == "" {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Token is required",
		})
		return
	}

	// TODO: Implement actual token verification logic
	// For now, return a mock response
	response := VerifyAccessTokenResponse{
		Valid:   true,
		Message: "Token is valid",
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// RenewAccessToken renews an access token using a refresh token
// @Summary Renew access token
// @Description Renews an access token using a valid refresh token (vrto)
// @Tags XFWToken
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body RenewAccessTokenRequest true "Refresh token request"
// @Success 200 {object} RenewAccessTokenResponse
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Router /api/v1/XFWToken/renew-access-token [post]
func RenewAccessToken(w http.ResponseWriter, r *http.Request) {
	var req RenewAccessTokenRequest

	// Decode request body
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "Invalid request body",
		})
		return
	}

	if req.Vrto == "" {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error": "vrto (refresh token) is required",
		})
		return
	}

	// TODO: Implement actual token renewal logic
	// For now, return a mock response
	response := RenewAccessTokenResponse{
		AccessToken:  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
		RefreshToken: req.Vrto,
		ExpiresIn:    3600,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}